// apps/web/src/lib/api/generated/models/validationExceptionResponse.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Formex
 * OpenAPI spec version: 0.0.1
 */
import type { ValidationExceptionResponseErrors } from "./validationExceptionResponseErrors";

export type ValidationExceptionResponse = {
  /** Errors overview. */
  message: string;
  /** A detailed description of each field that failed validation. */
  errors: ValidationExceptionResponseErrors;
};