// apps/web/src/lib/api/generated/student-feedback/student-feedback.zod.ts
export const StudentFeedbackLinkShowResponse = zod.object({
  "data": zod.object({
  "course_id": zod.string(),
  "semester": zod.number(),
  "academic_year": zod.number(),
  "feedback_type": zod.string(),
  "link_active": zod.boolean()
})
})

export const StudentFeedbackLinkStartParams = zod.object({
  "token": zod.string()
})

export const studentFeedbackLinkStartBodyRegisterNoMax = 100;



export const StudentFeedbackLinkStartBody = zod.object({
  "register_no": zod.string().max(studentFeedbackLinkStartBodyRegisterNoMax),
  "date_of_birth": zod.iso.datetime({})
})

export const studentFeedbackLinkStartResponseDataQuestionsOneMin = 0;
export const studentFeedbackLinkStartResponseDataQuestionsOneMax = 0;



export const StudentFeedbackLinkStartResponse = zod.object({
  "data": zod.object({
  "already_submitted": zod.boolean(),
  "student": zod.object({
  "id": zod.string(),
  "full_name": zod.string(),
  "register_no": zod.string().nullable()
}),
  "questions": zod.union([zod.array(zod.string()).min(studentFeedbackLinkStartResponseDataQuestionsOneMin).max(studentFeedbackLinkStartResponseDataQuestionsOneMax),zod.array(zod.object({
  "id": zod.string(),
  "institution_id": zod.string(),
  "question": zod.string(),
  "order_index": zod.number(),
  "created_at": zod.iso.datetime({}).nullable(),
  "updated_at": zod.iso.datetime({}).nullable()
}))])
})
})

export const StudentFeedbackLinkSubmitParams = zod.object({
  "token": zod.string()
})

export const studentFeedbackLinkSubmitBodyRegisterNoMax = 100;

export const studentFeedbackLinkSubmitBodyRatingsItemRatingMax = 5;




export const StudentFeedbackLinkSubmitBody = zod.object({
  "register_no": zod.string().max(studentFeedbackLinkSubmitBodyRegisterNoMax),
  "date_of_birth": zod.iso.datetime({}),
  "ratings": zod.array(zod.object({
  "question_id": zod.string(),
  "rating": zod.number().min(1).max(studentFeedbackLinkSubmitBodyRatingsItemRatingMax)
})).min(1)
})